Added funcitonality to remove btn
remove btn removes element from DOM
add-btn is disbled until removed from DOM

customers can only add one of each item to cart.

Realistyically, this would not be useful in a real world scenario, however, following the figma design, there is nothing about adding multiple of one item.

If this were a real world scanario, I could identify how many of an 'item' is in the order.items array and then render a 'x 3', if three were in the cart for example.